WIP on Client Errors tab updates
