+++ admin/tabs/activity/clienterrors/ClientErrorDetail.js
@@ -4,70 +4,57 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
-import {filler, table, tbody, td, th, tr} from '@xh/hoist/cmp/layout';
+import {div, filler, h3, hframe, table, tbody, td, th, tr} from '@xh/hoist/cmp/layout';
 import {hoistCmp} from '@xh/hoist/core';
-import {button} from '@xh/hoist/desktop/cmp/button';
 import {clipboardButton} from '@xh/hoist/desktop/cmp/clipboard';
-import {jsonInput} from '@xh/hoist/desktop/cmp/input';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
-import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {fmtDateTime} from '@xh/hoist/format';
-import {dialog} from '@xh/hoist/kit/blueprint';
+import {Icon} from '@xh/hoist/icon';
 
 export const clientErrorDetail = hoistCmp.factory(
     ({model}) => {
-        const rec = model.detailRecord;
+        const {selectedRecord, formattedErrorJson} = model;
+        if (!selectedRecord) return null;
 
-        if (!rec) return null;
-
-        return dialog({
-            title: 'Error Details',
-            style: {width: 1000},
-            isOpen: true,
-            onClose: () => model.closeDetail(),
-            items: detail()
-        });
-    }
-);
-
-const detail = hoistCmp.factory(
-    ({model}) => {
-        const {data} = model.detailRecord;
+        const {data} = selectedRecord;
         return panel({
-            items: [
-                table({
-                    className: 'xh-admin-error-detail',
+            className: 'xh-admin-client-errors__detail',
+            model: {
+                side: 'bottom',
+                defaultSize: 300
+            },
+            item: hframe(
+                table(
+                    tbody(
+                        tr(th('User:'), td(data.username)),
+                        tr(th('Message:'), td(data.msg || 'None provided')),
+                        tr(th('User Alerted:'), td(`${data.userAlerted}`)),
+                        tr(th('Device/Browser:'), td(`${data.device}/${data.browser}`)),
+                        tr(th('Agent:'), td(data.userAgent)),
+                        tr(th('App Version:'), td(data.appVersion)),
+                        tr(th('Environment:'), td(data.appEnvironment)),
+                        tr(th('Date:'), td(fmtDateTime(data.dateCreated)))
+                    )
+                ),
+                panel({
+                    flex: 1,
+                    className: 'xh-border-left',
                     items: [
-                        tbody(
-                            tr(th('User:'), td(data.username)),
-                            tr(th('Message:'), td(data.msg || 'None provided')),
-                            tr(th('User Alerted:'), td(`${data.userAlerted}`)),
-                            tr(th('Device/Browser:'), td(`${data.device}/${data.browser}`)),
-                            tr(th('Agent:'), td(data.userAgent)),
-                            tr(th('App Version:'), td(data.appVersion)),
-                            tr(th('Environment:'), td(data.appEnvironment)),
-                            tr(th('Date:'), td(fmtDateTime(data.dateCreated)))
-                        )
+                        h3(Icon.json(), 'Additional Data'),
+                        div({
+                            className: `xh-admin-activity-detail__json ${formattedErrorJson ? '' : 'xh-admin-activity-detail__json--empty'}`,
+                            item: formattedErrorJson ?? 'No additional details available.'
+                        })
+                    ],
+                    bbar: [
+                        filler(),
+                        clipboardButton({
+                            getCopyText: () => formattedErrorJson,
+                            successMessage: 'Error data copied to clipboard.',
+                            disabled: !formattedErrorJson,
+                            outlined: true
+                        })
                     ]
-                }),
-                jsonInput({
-                    value: data.error,
-                    disabled: true,
-                    height: 450,
-                    width: '100%',
-                    editorProps: {lineWrapping: true}
-                })
-            ],
-            bbar: toolbar(
-                filler(),
-                clipboardButton({
-                    getCopyText: () => data.error,
-                    successMessage: 'Error details copied to clipboard.'
-                }),
-                button({
-                    text: 'Close',
-                    intent: 'primary',
-                    onClick: () => model.closeDetail()
                 })
             )
         });
