+++ admin/tabs/activity/clienterrors/ClientErrorsView.js
@@ -6,30 +6,29 @@
  */
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
 import {filler} from '@xh/hoist/cmp/layout';
+import {storeFilterField} from '@xh/hoist/cmp/store';
 import {creates, hoistCmp} from '@xh/hoist/core';
-import {dimensionChooser} from '@xh/hoist/desktop/cmp/dimensionchooser';
-import {button, exportButton, refreshButton} from '@xh/hoist/desktop/cmp/button';
+import {button, colChooserButton, exportButton, refreshButton} from '@xh/hoist/desktop/cmp/button';
 import {dateInput, textInput} from '@xh/hoist/desktop/cmp/input';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbar, toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
 import {LocalDate} from '@xh/hoist/utils/datetime';
 import {clientErrorDetail} from './ClientErrorDetail';
-import {ClientErrorModel} from './ClientErrorModel';
+import {ClientErrorsModel} from './ClientErrorsModel';
 
-export const clientErrorPanel = hoistCmp.factory({
-    model: creates(ClientErrorModel),
+export const clientErrorsView = hoistCmp.factory({
+    model: creates(ClientErrorsModel),
 
     render({model}) {
         return panel({
-            mask: 'onLoad',
+            className: 'xh-admin-client-errors',
             tbar: tbar(),
             items: [
-                grid({
-                    onRowDoubleClicked: (e) => model.openDetail(e.data)
-                }),
+                grid(),
                 clientErrorDetail()
-            ]
+            ],
+            mask: 'onLoad'
         });
     }
 });
@@ -37,7 +36,6 @@ export const clientErrorPanel = hoistCmp.factory({
 const tbar = hoistCmp.factory(
     ({model}) => {
         return toolbar(
-            dimensionChooser(),
             button({
                 icon: Icon.angleLeft(),
                 onClick: () => model.adjustDates('subtract')
@@ -60,18 +58,12 @@ const tbar = hoistCmp.factory(
             refreshButton(),
             filler(),
             gridCountLabel({unit: 'error'}),
+            storeFilterField(),
+            colChooserButton(),
             exportButton()
         );
     }
 );
 
-const dateProps = {
-    popoverPosition: 'bottom',
-    valueType: 'localDate',
-    width: 120
-};
-
-const textProps = {
-    width: 150,
-    enableClear: true
-};
+const dateProps = {popoverPosition: 'bottom', valueType: 'localDate', width: 120};
+const textProps = {width: 150, enableClear: true};
